--- Tina/week1/ToDoListApp/src/components/add/TextInput.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import styled from 'styled-components/native';
-import THEME_COLOR from '../../styles/theme-color';
-import {useTask} from '../../store/task';
-
-const Input = () => {
-  const {task, setTask} = useTask();
-  return <TextInput value={task} onChangeText={e => setTask(e)} />;
-};
-
-export default Input;
-
-const TextInput = styled.TextInput`
-  height: 56px;
-  width: 330px;
-  border-radius: 20px;
-  padding: 16px 30px;
-  background-color: ${THEME_COLOR.white};
-`;
